+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### BioBlend v
 
-* Added support for Galaxy release 23.2.
+* Dropped support for Python 3.7. Added support for Python 3.12. Added support
+  for Galaxy release 23.2.
 
 * Parameters after ``password`` in the ``__init__()`` method of the
   ``GalaxyClient``, ``GalaxyInstance`` and ``ToolShedInstance`` classes are now